Created a dictionary that handles updates and new trains respectively. Successfully plotted one train
Much of the time was focused on creating a dictionary that makes working with the import data pieces easier. This dictionary allows you to instantiate new trains as well as update previous ones. Updating is based on either a train being new, at a station, or in between stations. More work has to be done type annotating, commenting/doc strings, actually plotting all points/running the api every 30 seconds.